+++ src/app/layout.js
@@ -2,7 +2,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import Header from "../components/header";
 import Footer from "../components/footer";
-import {AuthProvider} from "@/app/authContext";
+import {AuthProvider} from "@/app/AuthContext";
 import { Toaster } from "@/components/ui/toaster"
 
 const inter = Inter({ subsets: ["latin"] });
