+++ src/app/login/page.js
@@ -6,7 +6,7 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import {useState} from "react";
-import {useAuth} from "@/app/authContext";
+import {useAuth} from "@/app/AuthContext";
 import { useRouter } from "next/navigation";
 
 export const description =
